Addition of a lexing position limit, after which no new token can be created
The token started before the limit are lexed till their end, even
after the position limit.

Complementary, adding an "ignore last token" flag to the context
evaluator, to parse up to the las token, and keep the last token
only for prediction/completion in user code.

Bumping the package version to 1.1 due to new features.